Stub in basic example Tooltip compound component
