+++ package.json
@@ -130,6 +130,7 @@
     "@pollyjs/adapter": "^5.0.0",
     "@pollyjs/core": "^5.1.0",
     "@pollyjs/persister-fs": "^5.0.0",
+    "@react-types/tooltip": "^3.1.6",
     "@slack/web-api": "^5.10.0",
     "@sourcegraph/eslint-config": "^0.27.0",
     "@sourcegraph/eslint-plugin-sourcegraph": "1.0.0",
@@ -368,6 +369,7 @@
     "@reach/tabs": "^0.16.4",
     "@reach/visually-hidden": "^0.16.0",
     "@react-aria/live-announcer": "^3.0.5",
+    "@react-spectrum/tooltip": "^3.1.9",
     "@sentry/browser": "^6.13.2",
     "@slimsag/react-shortcuts": "^1.2.1",
     "@sourcegraph/extension-api-classes": "^1.1.0",
